Branch bugfix - changes for issue 11 (bugzilla 1131). Add restart capability When jobs analysis, forecast_step1, and forecast_step2 start - check if restart files exist in GESIN, and if so then restart from the latest set Upon getting a failure - save restart files to GESOUT - send FATAL message and exit with non-zero code Also required these changes - change frequency of restarts (6 hours for analysis, 12 for forecast) - change granularity of some script vars to YYYYMMDDHH as restarts can now be more often than daily

+++ scripts/exrtofs_glo_analysis.sh
@@ -11,6 +11,9 @@
 #           for the RTOFS_GLO Ocean model                                     #
 #                                                                             #
 # Sub-scripts called:                                                         #
+#   rtofs_runstaging.sh - gets input files from COMIN FIX and PARM            #
+#   rtofs_submit.sh - submits the RTOFS HYCOM simulation                      #
+#   rtofs_tmp2com.sh - copies the products to COMOUT                          #
 #                                                                             #
 # Script history log:                                                         #
 # 2010-07-30  Ilya Rivin                                                      #
@@ -26,61 +29,52 @@ msg="RTOFS_GLO_ANALYSIS JOB has begun on $(hostname) at $(date)"
 postmsg "$msg"
 
 # --------------------------------------------------------------------------- #
-# 0. date and time stuff
+# 0. set some defaults
 
-  export fcstdays=${fcstdays:-2}
-  export enddate=${analysis_end:-$PDY}
-  export startdate=$($NDATE -$(expr $fcstdays \* 24) ${enddate}${mycyc} | cut -c1-8)
-  export inputgrid=${inputgrid:-navy_0.08}
-
-# --------------------------------------------------------------------------- #
-# 1. Get input files
-  ksh ${USHrtofs}/${RUN}_runstaging.sh
+typeset -Z5 SSSSS
+export fcstdays=${fcstdays:-2}
+export inputgrid=${inputgrid:-navy_0.08}
 
 # --------------------------------------------------------------------------- #
-# 2. Get initial conditions (restart files)
-
-  typeset -Z5 sc
-  typeset -Z5 SSSSS
-
-  LEAD=$(expr $fcstdays \* -24)
-
-  HYCOMrestTplate=${RUN}_${modID}.t${mycyc}z.n${LEAD}.restart
-  CICErestTplate=${RUN}_${modID}.t${mycyc}z.n${LEAD}.restart_cice
+# 1. Get initial conditions (restart files)
 
   test -f restart_in.a && rm -f restart_in.a
   test -f restart_in.b && rm -f restart_in.b
 
+# loop through nowcast hours starting with most recent checking if restart files exist
+  restart_found=no
   if [ $RESTART = YES ]
   then
-     anal_hour=$(${NHOUR} ${enddate}${mycyc} ${PDY}${mycyc})
-     while [ $anal_hour -ge $LEAD ]
-     do       
-       if [ -s $GESIN/${RUN}_${modID}.t${mycyc}z.n${anal_hour}.restart}.b ]
+     # LEAD is the last forecast hour minus one
+     LEAD=$(expr $fcstdays \* -24 + 1)
+     for anal_hour in $(seq -w -1 -1 $LEAD)
+     do
+# find most recent analysis restart
+       if [ -s $GESIN/${RUN}_${modID}.t${mycyc}z.n${anal_hour}.restart.a -a \
+            -s $GESIN/${RUN}_${modID}.t${mycyc}z.n${anal_hour}.restart.b -a \
+            -s $GESIN/${RUN}_${modID}.t${mycyc}z.n${anal_hour}.restart_cice ]
        then
-         # find most recent analysis restart
          ln -s -f $GESIN/${RUN}_${modID}.t${mycyc}z.n${anal_hour}.restart.b restart_in.b
          ln -s -f $GESIN/${RUN}_${modID}.t${mycyc}z.n${anal_hour}.restart.a restart_in.a
          ln -s -f $GESIN/${RUN}_${modID}.t${mycyc}z.n${anal_hour}.restart_cice cice.restart_in
          echo "Analysis is started from restart: $GESIN/${RUN}_${modID}.t${mycyc}z.n${anal_hour}.restart.[ab]" \
-              " and ${RUN}_${modID}.t${mycyc}z.n${anal_hour}.restart_cice"
-
-       # replace limits file
-         sday=$($USHrtofs/rtofs_date_normal2hycom.sh $(${NDATE} +$anal_hour ${PDY}${mycyc}))
-         echo "  $sday $eday false false  " > limits
+              "and ${RUN}_${modID}.t${mycyc}z.n${anal_hour}.restart_cice"
+         restart_found=yes
          break
-
        fi
-
-       anal_hour=$(expr $anal_hour - 1)
      done
-     if [ $anal_hour -lt $LEAD ]
+
+     if [ $restart_found = no ]
      then
         $USHrtofs/${RUN}_abort.sh  "FATAL ERROR: $job Missing Restart File" \
           "No restart_in.[ab] or cice.restart_in in $GESIN" 2
-    fi 
+     fi
 
   else 
+    LEAD=$(expr $fcstdays \* -24)
+    HYCOMrestTplate=${RUN}_${modID}.t${mycyc}z.n${LEAD}.restart
+    CICErestTplate=${RUN}_${modID}.t${mycyc}z.n${LEAD}.restart_cice
+
     if [ -s ${COMIN}/${HYCOMrestTplate}.a ] && \
        [ -s ${COMIN}/${HYCOMrestTplate}.b ] && \
        [ -s ${COMIN}/${CICErestTplate} ] 
@@ -88,77 +82,125 @@ postmsg "$msg"
       ln -s -f ${COMIN}/${HYCOMrestTplate}.a restart_in.a
       ln -s -f ${COMIN}/${HYCOMrestTplate}.b restart_in.b
       ln -s -f ${COMIN}/${CICErestTplate} cice.restart_in
-      echo "Analysis is started from restart: ${COMIN}/${HYCOMrestTplate}.[ab] and ${COMIN}/${CICErestTplate}"
+      echo "Analysis is started from restart: ${COMIN}/${HYCOMrestTplate}.[ab] and ${CICErestTplate}"
     else
       $USHrtofs/${RUN}_abort.sh "FATAL ERROR: $job Missing Restart File" \
         "No restart_in.[ab] or cice.restart_in in $COMIN" 3
     fi
   fi
+
   echo './cice.restart_in' > cice.restart_file
 
   if [ -s restart_in.a -a -s restart_in.b -a -s cice.restart_in ]
   then
     echo "Initial restart files copied"
-    basetime=1900123100
-    hdate=$(awk '{  if (NR==2) { print $5 } }'  < restart_in.b | cut -d. -f1)
-    dater=$($NDATE $(expr $hdate \* 24)  ${basetime})
-    if [ ${dater} -ne ${startdate}${mycyc} ] 
-    then 
-      $USHrtofs/${RUN}_abort.sh  "FATAL ERROR: $job Mismatch Restart dates" \
-        "expected=${startdate}${mycyc} actual=$dater" 2
-    fi
+    export startdate=$(${USHrtofs}/rtofs_date4restart.sh restart_in.b)
+    export enddate=${analysis_end:-${PDY}${mycyc}}
   else
-      $USHrtofs/${RUN}_abort.sh "FATAL ERROR: $job Missing Restart File" \
+    $USHrtofs/${RUN}_abort.sh "FATAL ERROR: $job Missing Restart File" \
         "No restart_in.[ab] or cice.restart_in in $DATA" 911
   fi
 
 # --------------------------------------------------------------------------- #
-# 2. Run analysis
+# 2. get input files
+
+  ksh ${USHrtofs}/${RUN}_runstaging.sh
+
+# --------------------------------------------------------------------------- #
+# 3. Run analysis
 
   ${USHrtofs}/${RUN}_submit.sh
 
   ok="unknown"
   test -s ${DATA}/summary_out && ok=$(tail -1 ${DATA}/summary_out)
-  if [ "$ok" = "normal stop" ]; then
+  if [ "$ok" = "normal stop" ]
+  then
     modelstatus=0
   else
     modelstatus=1
-  fi 
-  if [ $modelstatus = 0 ]
-  then
-      # Copy to /com initial restart file for forecast step
-      for rfile in ${DATA}/restart_out.b ${DATA}/restart_out1.b 
-      do
-        cdate=$(${USHrtofs}/rtofs_date4restart.sh $rfile)
-        if [ $cdate -eq $PDY$mycyc ]
-        then          
-          YYYYMMDD=$(echo $cdate | cut -c1-8)
-          YYYYDDD=$(sh $USHutil/date2jday.sh $YYYYMMDD)
-          YYYY=$(echo $YYYYDDD | cut -c1-4)
-          DDD=$(echo $YYYYDDD | cut -c5-7)
-          MM=$(echo $YYYYMMDD | cut -c5-6)
-          DD=$(echo $YYYYMMDD | cut -c7-8)
-          HH=$(echo $cdate | cut -c9-10)
-          SSSSS=$(expr $HH \* 3600)
-          HYCOMrestTplate=${RUN}_${modID}.t${mycyc}z.n${HH}.restart
-          CICErestTplate=${RUN}_${modID}.t${mycyc}z.n${HH}.restart_cice
-          cp -p $rfile ${COMOUT}/${HYCOMrestTplate}.b
-          cp -p ${rfile%.b}.a ${COMOUT}/${HYCOMrestTplate}.a
-          cp -p cice.restart.${YYYY}-${MM}-${DD}-${SSSSS} ${COMOUT}/${CICErestTplate}
-        fi
-      done
-    echo "done" >$COMOUT/${RUN}_${modID}.t${mycyc}z.anal.log
-  else
-    $USHrtofs/${RUN}_abort.sh "FATAL ERROR: $job Abnormal model exit" \
-       "problem with analysis model run - return code $modelstatus" $modelstatus
   fi
 
 # --------------------------------------------------------------------------- #
-# 3. Copy results in the archive
+# 4. Copy results to com
 
+#
+# Note that on a failure, tmp2com will still copy archives after the latest restart
   if [ $SENDCOM = 'YES' ]
   then
     ksh ${USHrtofs}/${RUN}_tmp2com.sh
+  fi
+
+# --------------------------------------------------------------------------- #
+# 5. Check if run ran to completion and copy restart files to appropriate
+#    place depending on that status.
+
+# if model ran, then copy both restart files (n00, n-06) to comout
+#   n00 - for forecast step1
+#   n-06 - for PDYp1 incup step
+  if [ $modelstatus = 0 ]
+  then
+      for rfile in ${DATA}/restart_out.b ${DATA}/restart_out1.b 
+      do
+        cdate=$(${USHrtofs}/rtofs_date4restart.sh $rfile)
+        YYYY=$(echo $cdate | cut -c1-4)
+        MM=$(echo $cdate | cut -c5-6)
+        DD=$(echo $cdate | cut -c7-8)
+        HH=$(echo $cdate | cut -c9-10)
+        SSSSS=$(expr $HH \* 3600)
+        LEAD=$($NHOUR ${cdate} ${PDY}${mycyc})
+        HYCOMrestTplate=${RUN}_${modID}.t${mycyc}z.n${LEAD}.restart
+        CICErestTplate=${RUN}_${modID}.t${mycyc}z.n${LEAD}.restart_cice
+        cp -p $rfile ${COMOUT}/${HYCOMrestTplate}.b
+        cp -p ${rfile%.b}.a ${COMOUT}/${HYCOMrestTplate}.a
+        cp -p cice.restart.${YYYY}-${MM}-${DD}-${SSSSS} ${COMOUT}/${CICErestTplate}
+      done
+    echo "done" >$COMOUT/${RUN}_${modID}.t${mycyc}z.anal.log
+  else
+    date_out=0 ; date_out1=0; rcount=0
+    if [ -s ${DATA}/restart_out.b ]
+    then
+      date_out=$(${USHrtofs}/rtofs_date4restart.sh ${DATA}/restart_out.b)
+      let rcount=rcount+1
+    fi
+    if [ -s ${DATA}/restart_out1.b ]
+    then
+      date_out1=$(${USHrtofs}/rtofs_date4restart.sh ${DATA}/restart_out1.b)
+      let rcount=rcount+1
+    fi
+# copy latest restart file (if found) to GESOUT
+    if [ $rcount -ne 0 ]
+    then
+      if [ ${date_out} -gt ${date_out1} ]
+      then
+        rfile=${DATA}/restart_out.b
+        cdate=${date_out}
+      else
+        rfile=${DATA}/restart_out1.b
+        cdate=${date_out1}
+      fi
+      YYYY=$(echo $cdate | cut -c1-4)
+      MM=$(echo $cdate | cut -c5-6)
+      DD=$(echo $cdate | cut -c7-8)
+      HH=$(echo $cdate | cut -c9-10)
+      SSSSS=$(expr $HH \* 3600)
+      LEAD=$($NHOUR ${cdate} ${PDY}${mycyc})
+      HYCOMrestTplate=${RUN}_${modID}.t${mycyc}z.n${LEAD}.restart
+      CICErestTplate=${RUN}_${modID}.t${mycyc}z.n${LEAD}.restart_cice
+      cp -p $rfile ${GESOUT}/${HYCOMrestTplate}.b
+      cp -p ${rfile%.b}.a ${GESOUT}/${HYCOMrestTplate}.a
+      cp -p cice.restart.${YYYY}-${MM}-${DD}-${SSSSS} ${GESOUT}/${CICErestTplate}
+# check to see if LEAD is -06 (current incremental update length)
+# if it is, then  need to also copy this restart file to COMOUT 
+# note -06 is same as inc_hours in exrtofs_glo_incup.sh
+      if [ $LEAD -eq -06 ]
+      then
+        cp -p $rfile ${COMOUT}/${HYCOMrestTplate}.b
+        cp -p ${rfile%.b}.a ${COMOUT}/${HYCOMrestTplate}.a
+        cp -p cice.restart.${YYYY}-${MM}-${DD}-${SSSSS} ${COMOUT}/${CICErestTplate}
+      fi
+    fi
+    $USHrtofs/${RUN}_abort.sh "FATAL ERROR: $job Abnormal model exit" \
+       "problem with analysis model run - return code $modelstatus" $modelstatus
   fi
 
 #################################################
